# api-docs.yaml
openapi: 3.0.3
info:
  title: Evidence And Cyber Backend API
  version: 0.1.0
  description: API for endpoint inventory, agent ingestion, detection management, remote help and command orchestration.
servers:
  - url: /api/v1
paths:
...
  /users:
    get:
      tags: [Directory]
      summary: Get all users
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
  /users/{id}:
    get:
      tags: [Directory]
      summary: Get user by id
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
  /users/{id}/roles:
    put:
      tags: [Directory]
      summary: Replace user role assignments
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleAssignmentRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
  /roles:
    get:
      tags: [Directory]
      summary: Get all roles
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleDto'
  /permissions:
    get:
      tags: [Directory]
      summary: Get all permissions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Permission list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionDto'
  /roles/{id}/permissions:
    put:
      tags: [Directory]
      summary: Replace role permission assignments
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePermissionAssignmentRequest'
      responses:
        '200':
          description: Updated role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDto'
...
  /remote-help-requests:
    get:
      tags: [RemoteSupport]
      summary: Get all remote help requests
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Remote help request list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RemoteHelpRequestDto'
  /remote-help-requests/{id}:
    get:
      tags: [RemoteSupport]
      summary: Get remote help request detail
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200':
          description: Help request detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteHelpRequestDto'
  /remote-help-requests/{id}/accept:
    post:
      tags: [RemoteSupport]
      summary: Accept remote help request
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200':
          description: Accepted help request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteHelpRequestDto'
  /remote-sessions:
    get:
      tags: [RemoteSupport]
      summary: Get remote sessions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Remote session list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RemoteSessionDto'
    post:
      tags: [RemoteSupport]
      summary: Create remote session
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoteSessionCreateRequest'
      responses:
        '200':
          description: Created remote session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteSessionDto'
  /remote-sessions/{id}/approvals:
    post:
      tags: [RemoteSupport]
      summary: Create control approval record
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlApprovalRequest'
      responses:
        '200':
          description: Created control approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlApprovalDto'
...
components:
  parameters:
    UuidPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRoleAssignmentRequest:
      type: object
      properties:
        roleIds:
          type: array
          items:
            type: string
            format: uuid
    RolePermissionAssignmentRequest:
      type: object
      properties:
        permissionIds:
          type: array
          items:
            type: string
            format: uuid
    PermissionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
    RoleDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        system:
          type: boolean
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionDto'
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        adUsername:
          type: string
        displayName:
          type: string
        email:
          type: string
        department:
          type: string
        enabled:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        source:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'
...
    RemoteHelpRequestDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        requestedByUsername:
          type: string
        requestedByDisplayName:
          type: string
        message:
          type: string
        status:
          type: string
        requestedAt:
          type: string
          format: date-time
        acceptedByUserId:
          type: string
          format: uuid
        acceptedAt:
          type: string
          format: date-time
    RemoteSessionCreateRequest:
      type: object
      properties:
        helpRequestId:
          type: string
          format: uuid
        sessionType:
          type: string
        provider:
          type: string
    RemoteSessionDto:
      allOf:
        - $ref: '#/components/schemas/RemoteSessionCreateRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            adminUserId:
              type: string
              format: uuid
            status:
              type: string
            startedAt:
              type: string
              format: date-time
            endedAt:
              type: string
              format: date-time
    ControlApprovalRequest:
      type: object
      properties:
        decision:
          type: string
        decidedByUsername:
          type: string
        note:
          type: string
    ControlApprovalDto:
      allOf:
        - $ref: '#/components/schemas/ControlApprovalRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            remoteSessionId:
              type: string
              format: uuid
            requestedAt:
              type: string
              format: date-time
            decidedAt:
              type: string
              format: date-time
...